fix: pnpm build
