+++ app/providers.tsx
@@ -7,8 +7,6 @@ import { RainbowKitProvider } from "@rainbow-me/rainbowkit";
 import "@rainbow-me/rainbowkit/styles.css";
 import { config } from "@/config/wagmi";
 import { useState } from "react";
-import { WalletConnectorProvider } from "@/components/providers/WalletConnectorProvider";
-import { StakingServiceProvider } from "@/components/providers/StakingServiceProvider";
 
 export function Providers({ children }: { children: React.ReactNode }) {
   const [queryClient] = useState(() => new QueryClient());
