--- components/modals/DelegateModal.tsx
@@ -1,7 +1,0 @@
-export function DelegateModal() {
-  return (
-    <div>
-      <h1>Delegate Modal</h1>
-    </div>
-  );
-}
