--- components/modals/UndelegateModal.tsx
@@ -1,7 +1,0 @@
-export function UndelegateModal() {
-  return (
-    <div>
-      <h1>Undelegate Modal</h1>
-    </div>
-  );
-}
