--- components/modals/WithdrawModal.tsx
@@ -1,7 +1,0 @@
-export function WithdrawModal() {
-  return (
-    <div>
-      <h1>Withdraw Modal</h1>
-    </div>
-  );
-}
