+++ components/providers/OperatorsProvider.tsx
@@ -1,6 +1,5 @@
-import { createContext, useContext, ReactNode } from "react";
+import { ReactNode } from "react";
 import { useOperators } from "@/hooks/useOperators";
-import { OperatorInfo } from "@/types/operator";
 import { OperatorsContext } from "@/contexts/OperatorsContext";
 
 export function OperatorsProvider({ children }: { children: ReactNode }) {
